backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "jean Bertholat",email = "jean.bertholat@alten.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"

package-mode = false

dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "supabase (>=2.17.0,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "injector (>=0.22.0,<1.0.0)"
]

packages = [
    { include = "backend", from = "src/"},
]

[tool.poetry.scripts]
start = "backend.app:startup"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
